DERBY-5357 SQLJ.INSTALL_JAR shouldn't use identifier as file name
Since SQL identifiers can contain arbitrary characters, it is not safe
to use them as is as part of a file name. Trying ot map parts of the
name by excluding unsafe characters leads to a chance of name
collision. So, we have changed the naming altogether.

This patch, derby-5357-with-tests-4, changes the name (and location)
of the jar files stored in a database. The name is now based on UUID,
and no subdirectories under the directory "jar" are used: all
jar-files reside directly in the "jar" database directory, and the
name is of the form

<Derby uuid string>[.]jar[.]G[0-9]+

where <Derby uuid string> has the form hhhhhhhh-hhhh-hhhh-hhhh-hhhhhhhhhhhh
where h id a lower case hex digit, and the suffix ".G[0-9]+" is the
version number as before.

The format is changed on hard upgrade, cf tests in Changes10_9.

Also, dblook has been updated to cater for this change.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@1302836 13f79535-47bb-0310-9956-ffa450edef68

+++ java/engine/org/apache/derby/impl/store/raw/data/RFResource.java
@@ -148,7 +148,36 @@ class RFResource implements FileResource {
 		return generationId;
 	}
 
-	/**
+    /**
+     * @see FileResource#removeJarDir
+     */
+    public void removeJarDir(String f) throws StandardException {
+        if (factory.isReadOnly())
+            throw StandardException.newException(SQLState.FILE_READ_ONLY);
+
+        ContextManager cm =
+            ContextService.getFactory().getCurrentContextManager();
+
+        RawTransaction tran =
+            factory.getRawStoreFactory().getXactFactory().findUserTransaction(
+                        factory.getRawStoreFactory(),
+                        cm,
+                        AccessFactoryGlobals.USER_TRANS_NAME);
+
+        StorageFile ff = factory.storageFactory.newStorageFile(f);
+        Serviceable s = new RemoveFile(ff);
+
+        // Since this code is only used during upgrade to post-10.8 databases
+        // we do no bother to build code for a special RemoveDirOperation and
+        // do tran.logAndDo (cf. logic in #remove). If the post-commit removal
+        // doesn't get completed, that is no big issue, the dirs can be removed
+        // by hand if need be. A prudent DBA will rerun the upgrade from a
+        // backup if something crashes anyway..
+
+        tran.addPostCommitWork(s);
+    }
+
+    /**
 	  @see FileResource#remove
 	  @exception StandardException Oops
 	  */
@@ -256,9 +285,16 @@ final class RemoveFile implements Serviceable, PrivilegedExceptionAction
     public Object run() throws StandardException {
         // SECURITY PERMISSION - MP1, OP5
         if (fileToGo.exists()) {
-            if (!fileToGo.delete()) {
-                throw StandardException.newException(
-                        SQLState.FILE_CANNOT_REMOVE_FILE, fileToGo);
+            if (fileToGo.isDirectory()) {
+                if (!fileToGo.deleteAll()) {
+                    throw StandardException.newException(
+                            SQLState.FILE_CANNOT_REMOVE_FILE, fileToGo);
+                }
+            } else {
+                if (!fileToGo.delete()) {
+                    throw StandardException.newException(
+                            SQLState.FILE_CANNOT_REMOVE_FILE, fileToGo);
+                }
             }
         }
         return null;
